Send PRs to both wpt and homebrew-cask-versions
Fixes https://github.com/foolip/safari-technology-preview-updater/issues/37.
