+++ README.md
@@ -1,6 +1,7 @@
 # Safari Technology Preview Cask Updater
 
-Scrapes [Downloads for Safari](https://developer.apple.com/safari/download/) and sends pull requests to [homebrew-cask-versions](https://github.com/Homebrew/homebrew-cask-versions).
+Scrapes [Downloads for Safari](https://developer.apple.com/safari/download/) and sends pull requests to [homebrew-cask-versions](https://github.com/Homebrew/homebrew-cask-versions) and [web-platform-tests](https://github.com/web-platform-tests/wpt).
 
 * [Azure Pipelines builds](https://dev.azure.com/foolip/safari-technology-preview-updater/_build?definitionId=6)
 * [homebrew-cask-versions pull requests](https://github.com/Homebrew/homebrew-cask-versions/pulls/autofoolip)
+* [web-platform-tests pull requests](https://github.com/web-platform-tests/wpt/pulls?utf8=%E2%9C%93&q=is%3Apr+author%3Aautofoolip+label%3A%22Azure+Pipelines%22)
